+++ src/metadata.jl
@@ -8,7 +8,7 @@ end
 
 using FlatBuffers
 
-@enum(Type_, BOOL = 0, INT8 = 1, INT16 = 2, INT32 = 3, INT64 = 4,
+@enum(DType, BOOL = 0, INT8 = 1, INT16 = 2, INT32 = 3, INT64 = 4,
   UINT8 = 5, UINT16 = 6, UINT32 = 7, UINT64 = 8,
   FLOAT = 9, DOUBLE = 10,  UTF8 = 11,  BINARY = 12,
   CATEGORY = 13, TIMESTAMP = 14, DATE = 15, TIME = 16)
@@ -20,7 +20,7 @@ using FlatBuffers
 # FlatBuffers.enumsizeof(::Type{TimeUnit}) = UInt8
 
 mutable struct PrimitiveArray
-    type_::Type_
+    dtype::DType
     encoding::Encoding
     offset::Int64
     length::Int64
@@ -28,6 +28,7 @@ mutable struct PrimitiveArray
     total_bytes::Int64
 end
 
+# TODO why are these done this way rather with an abstract type???
 mutable struct CategoryMetadata
     levels::PrimitiveArray
     ordered::Bool
@@ -57,8 +58,10 @@ mutable struct Column
     user_metadata::String
 end
 
-function Column(name::String, values::PrimitiveArray, metadata::TypeMetadata=nothing, user_metadata::String="")
-    return Column(name, values, FlatBuffers.typeorder(TypeMetadata, typeof(metadata)), metadata, user_metadata)
+function Column(name::String, values::PrimitiveArray, metadata::TypeMetadata=nothing,
+                user_metadata::String="")
+    Column(name, values, FlatBuffers.typeorder(TypeMetadata, typeof(metadata)),
+           metadata, user_metadata)
 end
 
 mutable struct CTable
@@ -73,7 +76,7 @@ end # module
 
 # wesm/feather/cpp/src/metadata_generated.h
 # wesm/feather/cpp/src/types.h
-const Type_2julia = Dict{Metadata.Type_,DataType}(
+const JULIA_TYPE_DICT = Dict{Metadata.DType,DataType}(
     Metadata.BOOL      => Arrow.Bool,
     Metadata.INT8      => Int8,
     Metadata.INT16     => Int16,
@@ -93,7 +96,7 @@ const Type_2julia = Dict{Metadata.Type_,DataType}(
     Metadata.TIME      => Int64
 )
 
-const julia2Type_ = Dict{DataType,Metadata.Type_}(
+const MDATA_TYPE_DICT = Dict{DataType,Metadata.DType}(
     Bool    => Metadata.BOOL,
     Int8    => Metadata.INT8,
     Int16   => Metadata.INT16,
@@ -114,10 +117,41 @@ const julia2Type_ = Dict{DataType,Metadata.Type_}(
 
 const NON_PRIMITIVE_TYPES = Set([Metadata.UTF8, Metadata.BINARY])
 
-const TimeUnit2julia = Dict{Metadata.TimeUnit,DataType}(
-    Metadata.SECOND => Arrow.Second,
-    Metadata.MILLISECOND => Arrow.Millisecond,
-    Metadata.MICROSECOND => Arrow.Microsecond,
-    Metadata.NANOSECOND => Arrow.Nanosecond
-)
-const julia2TimeUnit = Dict{DataType,Metadata.TimeUnit}([(v, k) for (k,v) in TimeUnit2julia])
+# TODO why is this a separate dict??
+# const JULIA_TIME_DICT = Dict{Metadata.TimeUnit,DataType}(
+#     Metadata.SECOND => Arrow.Second,
+#     Metadata.MILLISECOND => Arrow.Millisecond,
+#     Metadata.MICROSECOND => Arrow.Microsecond,
+#     Metadata.NANOSECOND => Arrow.Nanosecond
+# )
+# const MDATA_TIME_DICT = Dict{DataType,Metadata.TimeUnit}(v=>k for (k,v) in JULIA_TIME_DICT)
+
+
+# TODO this really doesn't seem like an ideal way to deal with this
+isprimitivetype(t::Metadata.DType) = t ∉ NON_PRIMITIVE_TYPES
+
+
+juliastoragetype(meta::Void, values_type::Metadata.DType) = JULIA_TYPE_DICT[values_type]
+function juliastoragetype(meta::Metadata.CategoryMetadata, values_type::Metadata.DType)
+    CategoricalString{JULIA_TYPE_DICT[values_type]}
+end
+function juliastoragetype(meta::Metadata.TimestampMetadata, values_type::Metadata.DType)
+    throw(ErrorException("Not implemented."))
+    # TODO define this!
+end
+
+# TODO finish implementing these
+juliatype(::Type{T}) where T = T
+
+# TODO these functions should be able to be combined with previous
+function schematype(::Type{T}, nullcount::Integer, nullable::Bool, wrs::Bool) where T
+    (nullcount == 0 && !nullable) ? T : Union{T,Missing}
+end
+function schematype(::Type{<:AbstractString}, nullcount::Integer, nullable::Bool, wrs::Bool)
+    s = wrs ? WeakRefString{UInt8} : String
+    (nullcount == 0 && !nullable) ? s : Union{s, Missing}
+end
+function schematype(::Type{CategoricalString{R}}, nullcount::Integer, nullable::Bool,
+                    wrs::Bool) where R
+    (nullcount == 0 && !nullable) ? CategoricalString{R} : Union{CategoricalString{R}, Missing}
+end
