+++ test/sandbox.jl
@@ -1,0 +1,9 @@
+using DataFrames
+using DataStreams
+using Feather
+
+
+s = Feather.Source("stringtest1.feather")
+
+v = Feather.constructcolumn(s, 1)
+
